requirements.txt
# Requirements pour OAPET Backend
# Installation: pip install -r requirements.txt

# Django et DRF
Django==5.1.7
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.3.0

# Base de données
psycopg2-binary==2.9.9
django-redis==5.4.0

# Authentification et sécurité
djangorestframework-simplejwt==5.3.0
django-allauth==0.63.6

# Configuration et environnement
python-dotenv==1.0.1
python-decouple==3.8

# Machine Learning
scikit-learn==1.3.0
xgboost==2.0.3
pandas==2.2.2
numpy==1.26.4
networkx==3.3
joblib==1.4.2

# Utilitaires
requests==2.31.0
python-dateutil==2.9.0.post0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring et logs
django-extensions==3.2.3
django-debug-toolbar==4.4.6

# Tests
pytest==8.2.2
pytest-django==4.8.0
factory-boy==3.3.0

# API Documentation
drf-spectacular==0.27.2

# Export et import
openpyxl==3.1.2
reportlab==4.2.0

# Tâches asynchrones (optionnel)
celery==5.3.6
redis==5.0.7

# Géolocalisation (optionnel)
geopy==2.4.1

# Validation et nettoyage des données
bleach==6.1.0
django-cleanup==8.0.0